+++ src/valency_anndata/datasets/_load_aufstehen.py
@@ -1,5 +1,4 @@
 from typing import Optional
-from huggingface_hub import snapshot_download
 import valency_anndata as val
 
 def aufstehen(
@@ -33,13 +32,9 @@ def aufstehen(
     more information about how the data was collected can be found at the
     following link: <https://pol.is/report/r6xd526vyjyjrj9navxrj>
     """
-    export_dir = snapshot_download(
-        repo_id="patcon/polis-aufstehen-2018",
-        repo_type="dataset",
-        # Suppress HF_TOKEN warning.
-        token=False,
+    adata = val.datasets.polis.load(
+        source="huggingface:patcon/polis-aufstehen-2018",
+        translate_to=translate_to,
     )
-    adata = val.datasets.polis.load(source=export_dir, translate_to=translate_to)
 
     return adata
-